refactor: update demos
