+++ packages/ui/src/components/va-form/VaForm.demo.vue
@@ -151,35 +151,6 @@
         </va-button>
       </va-form>
     </VbCard>
-
-    <VbCard title="reset form">
-      <va-button @click="$refs.resetForm.reset()">Reset form</va-button>
-      <va-form ref="resetForm">
-        <va-date-input
-          v-model="resetForm.date"
-          :rules="resetForm.validationRules"
-        />
-        <va-input
-          v-model="resetForm.input"
-          :rules="resetForm.validationRules"
-        />
-        <va-option-list
-          type="radio"
-          v-model="resetForm.optionsListValue"
-          :options="resetForm.options"
-          :rules="resetForm.validationRules"
-        />
-        <va-select
-          v-model="resetForm.select"
-          :options="resetForm.options"
-          :rules="resetForm.validationRules"
-        />
-        <va-time-input
-          v-model="resetForm.time"
-          :rule="resetForm.validationRules"
-        />
-      </va-form>
-    </VbCard>
   </VbDemo>
 </template>
 
@@ -190,10 +161,6 @@ import { VaSelect } from '../va-select'
 import { VaButton } from '../va-button'
 import VaFormReset from './VaForm-reset'
 import { VaDateInput } from '../va-date-input'
-import { VaTimeInput } from '../va-time-input'
-import { VaOptionList } from '../va-option-list'
-
-const OPTIONS = ['One', 'Two', 'Three']
 
 export default {
   components: {
@@ -203,8 +170,6 @@ export default {
     VaSelect,
     VaButton,
     VaDateInput,
-    VaTimeInput,
-    VaOptionList,
   },
   data () {
     return {
@@ -225,16 +190,6 @@ export default {
       },
       username: '',
       password: '',
-
-      resetForm: {
-        input: 'value',
-        date: new Date(),
-        time: new Date(),
-        options: OPTIONS,
-        select: OPTIONS[0],
-        optionsListValue: OPTIONS[0],
-        validationRules: [(value) => !!value || 'Field is required'],
-      },
     }
   },
   methods: {
